Wrap smart_ptr_t to portable type
